+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: minskim2 <minskim2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/03 14:46:49 by minskim2          #+#    #+#             */
-/*   Updated: 2021/05/03 22:34:36 by minskim2         ###   ########.fr       */
+/*   Updated: 2021/05/07 22:29:28 by minskim2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *dest, int fillchar, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *dest, int fillchar, size_t len)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*temp;
 
 	i = 0;
